refactored locator, map, partners, partnerslist, into their own component files, css styling, added links to customerpage, still needs mobile-friendly styling
